extract understack-flavor-matcher module
This will allow us to reuse the same code betwen the custom Argo
workflows for enrollment as well as Ironic hooks/inspection interfaces.
Eventually when we fully migrate to using ironic hooks, this can be
inlined again.
