+++ containers/ironic/Dockerfile.ironic
@@ -10,4 +10,5 @@ RUN apt-get update && \
     && apt-get clean && rm -rf /var/lib/apt/lists/*
 
 COPY python/ironic-understack /tmp/ironic-understack
+COPY python/understack-flavor-matcher /tmp/ironic-understack
 RUN /var/lib/openstack/bin/python -m pip install --no-cache --no-cache-dir /tmp/ironic-understack sushy-oem-idrac==6.0.0
