+++ python/understack-flavor-matcher/README.md
@@ -1,0 +1,1 @@
+Mini library to identify a hardware flavor of a machine.
