+++ python/understack-flavor-matcher/tests/test_matcher.py
@@ -1,7 +1,7 @@
 import pytest
-from ironic_understack.flavor_spec import FlavorSpec
-from ironic_understack.machine import Machine
-from ironic_understack.matcher import Matcher
+from flavor_matcher.flavor_spec import FlavorSpec
+from flavor_matcher.machine import Machine
+from flavor_matcher.matcher import Matcher
 
 
 @pytest.fixture
